refactor: component → own module
