+++ code/system/sys.net.webrtc/src/ui/ui.Connect/ui.Connect.tsx
@@ -1,6 +1,6 @@
-import { PeerInput } from '../ui.PeerInput';
 import { WebRtcInfo } from '../ui.Info';
-import { COLORS, Color, DEFAULTS, css, type t } from './common';
+import { DEFAULTS, css, type t } from './common';
+import { ConnectInput } from './ui.Connect.Input';
 
 export const Connect: React.FC<t.ConnectProps> = (props) => {
   const { fields = DEFAULTS.fields } = props;
@@ -17,8 +17,8 @@ export const Connect: React.FC<t.ConnectProps> = (props) => {
     }),
   };
 
-  const elTop = Wrangle.ConnectComponent(props, 'Top');
-  const elBottom = Wrangle.ConnectComponent(props, 'Bottom');
+  const elTop = <ConnectInput {...props} targetEdge={'Top'} />;
+  const elBottom = <ConnectInput {...props} targetEdge={'Bottom'} />;
 
   const elInfo = (
     <WebRtcInfo
@@ -51,44 +51,5 @@ const Wrangle = {
         bottom: edge === 'Bottom',
       },
     };
-  },
-
-  ConnectComponent(props: t.ConnectProps, targetEdge: t.VEdge) {
-    const data = props.info?.connect;
-    const self = data?.self;
-    const { edge = DEFAULTS.edge, card = DEFAULTS.card } = props;
-
-    if (!data || !self) return null;
-    if (edge !== targetEdge) return null;
-
-    const is = Wrangle.is(props);
-    const border = `solid 1px ${Color.alpha(COLORS.DARK, 0.1)}`;
-    const margin = card ? 25 : 15;
-    const styles = {
-      base: css({ boxSizing: 'border-box' }),
-      top: css({
-        borderBottom: border,
-        marginBottom: margin,
-      }),
-      bottom: css({
-        borderTop: border,
-        marginTop: margin,
-      }),
-    };
-
-    return (
-      <div {...css(styles.base, is.edge.top && styles.top, is.edge.bottom && styles.bottom)}>
-        <PeerInput
-          self={self}
-          remote={data.remote}
-          fields={['Peer:Self', 'Peer:Remote']}
-          spinning={data.spinning}
-          copiedMessage={props.copiedMessage}
-          onLocalCopied={data.onLocalCopied}
-          onRemoteChanged={data.onRemoteChanged}
-          onConnectRequest={data.onConnectRequest}
-        />
-      </div>
-    );
   },
 };
